pyproject.toml
[project]
name = "stt-tts-uz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "librosa>=0.10.0",
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
    "onnxruntime-gpu>=1.17.0",
    "optimum[onnxruntime-gpu]>=1.18.0",
    "phonemizer>=3.2.1",
    "scipy>=1.11.0",
    "torch>=2.0.0",
    "transformers>=4.38.0",
    "unidecode>=1.3.8",
]
